+++ src/components/ui/CheckboxDropdown.tsx
@@ -1,11 +1,11 @@
 import { useEffect, useMemo, useState } from "react";
-import { Button } from "./button";
+import { Button } from "./shad-cn/button";
 import {
   DropdownMenu,
   DropdownMenuCheckboxItem,
   DropdownMenuContent,
   DropdownMenuTrigger,
-} from "./dropdown-menu";
+} from "./shad-cn/dropdown-menu";
 
 type CheckboxDropdownProps<T> = {
   title: string;
