+++ src/components/ui/DateTimePopover.tsx
@@ -4,14 +4,14 @@ import {
   Popover,
   PopoverContent,
   PopoverTrigger,
-} from "~/components/ui/popover";
+} from "~/components/ui/shad-cn/popover";
 
 import { useEffect, useState } from "react";
 import { cn } from "src/utils";
 import { TABLE_DATE_FORMAT } from "~/constants";
-import { Button } from "./button";
-import { Calendar } from "./calendar";
-import { Input } from "./input";
+import { Button } from "./shad-cn/button";
+import { Calendar } from "./shad-cn/calendar";
+import { Input } from "./shad-cn/input";
 
 type DateTimePopoverProps = {
   onChange: (date: Date) => void;
